requirements (1).txt
google-generativeai
python-docx
markdown